refactor(web): componente seção de formulário Horários disponíveis
