+++ web/src/components/ClassScheduleForm/styles.css
@@ -1,0 +1,38 @@
+.schedule-item {
+  padding-bottom: 2.75rem;
+  border-bottom: 1px solid var(--color-line-in-white);
+
+  position: relative;
+}
+
+.schedule-item + .schedule-item {
+  margin-top: 3.65rem;
+}
+
+.interval-section {
+  display: flex;
+  column-gap: 1.6rem;
+}
+
+.delete-class-schedule-button {
+  color: var(--color-danger);
+
+  padding-inline: 2.5rem;
+  border: 0;
+
+  background: var(--color-title-in-primary);
+  cursor: pointer;
+
+  position: absolute;
+  bottom: 0;
+  left: 50%;
+  transform: translateX(-50%) translateY(50%);
+}
+
+@media (min-width: 700px) {
+  .schedule-section {
+    display: grid;
+    grid-template-columns: 2fr 1fr;
+    column-gap: 1.6rem;
+  }
+}
