+++ web/src/pages/MyProfile/styles.css
@@ -25,32 +25,6 @@
   /* margin-bottom: 4rem; */
 }
 
-#page-my-profile .schedule-item {
-  padding-bottom: 2.75rem;
-  border-bottom: 1px solid var(--color-line-in-white);
-
-  position: relative;
-}
-
-#page-my-profile .delete-class-schedule-button {
-  color: var(--color-danger);
-
-  padding-inline: 2.5rem;
-  border: 0;
-
-  background: var(--color-title-in-primary);
-  cursor: pointer;
-
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%) translateY(50%);
-}
-
-#page-my-profile .schedule-item + .schedule-item {
-  margin-top: 3.65rem;
-}
-
 @media (min-width: 700px) {
   #page-my-profile .header-content {
     padding-bottom: 12.8rem;
@@ -85,11 +59,5 @@
     display: grid;
     grid-template-columns: 2fr 1fr;
     column-gap: 3.2rem;
-  }
-
-  #page-my-profile fieldset .schedule-section {
-    display: grid;
-    grid-template-columns: 2fr 1fr 1fr;
-    column-gap: 1.6rem;
   }
 }
